AndroidApp: Unify signed / unsigned BE16 reads.

+++ AndroidApp/app/src/main/java/cz/xoft/deskemes/MuxChannelZero.java
@@ -148,7 +148,7 @@ class MuxChannelZero
 		short id = 0;
 		try
 		{
-			id = (short)aBar.readBE16();
+			id = aBar.readBE16();
 		}
 		catch (ByteArrayReader.DataEndReachedException exc)
 		{
